requirements.txt
alpaca-py==0.43.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.129.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
psycopg2==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
sentry-sdk==2.53.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.47
sqlmodel==0.0.37
sseclient-py==1.9.0
starlette==0.52.1
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
